refactor: move transaction update to transaction domain
