+++ src/modules/transaction/hooks/queries/useTransactions.js
@@ -5,15 +5,13 @@ import { LIMIT as limit, API_VERSION } from 'src/const/config';
 import { useCustomInfiniteQuery } from 'src/modules/common/hooks';
 // import client from 'src/utils/api/client';
 
-export const useTransactionsConfig =
-  () =>
-  (customConfig = {}) => ({
-    url: `/api/${API_VERSION}/transactions`,
-    method: 'get',
-    event: 'update.transactions',
-    ...customConfig,
-    params: { limit, ...(customConfig?.params || {}) },
-  });
+export const useTransactionsConfig = (customConfig = {}) => ({
+  url: `/api/${API_VERSION}/transactions`,
+  method: 'get',
+  event: 'update.transactions',
+  ...customConfig,
+  params: { limit, ...(customConfig?.params || {}) },
+});
 
 /**
  * Creates a custom hook for transaction list query
@@ -51,7 +49,7 @@ export const useTransactionsConfig =
 export const useTransactions = ({ config: customConfig = {}, options } = {}) => {
   const [hasUpdate, setHasUpdate] = useState(false);
   const queryClient = useQueryClient();
-  const config = useTransactionsConfig()(customConfig);
+  const config = useTransactionsConfig(customConfig);
 
   /* istanbul ignore next */
   // client.socket.on('new.transactions', () => {
