+++ src/modules/transaction/hooks/useTransactionUpdate.js
@@ -1,47 +1,17 @@
-import React, { useEffect } from 'react';
-import { toast } from 'react-toastify';
+import { useEffect } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
 import client from 'src/utils/api/client';
 import { TRANSACTIONS, AUTH } from 'src/const/queries';
-import { filterIncomingTransactions } from '@transaction/utils/helpers';
-import { fromRawLsk } from '@token/fungible/utils/lsk';
-import i18n from 'src/utils/i18n/i18n';
 import { useCurrentAccount } from '@account/hooks';
 import { useCurrentApplication } from '@blockchainApplication/manage/hooks';
 import { useTransactions, useTransactionsConfig } from '@transaction/hooks/queries';
+import {showNotificationsForIncomingTransactions} from '../utils'
 
-const showNotificationsForIncomingTransactions = (transactions, account, token) => {
-  filterIncomingTransactions(transactions, account).forEach((transaction) => {
-    const amount = fromRawLsk(transaction.params.amount);
-    if (amount > 0) {
-      const message = transaction.params.data
-        ? i18n.t('with message {{message}}', { message: transaction.params.data })
-        : '';
-      // @TODO: To be fixed in #4652
-      // eslint-disable-next-line no-new
-      new Notification(i18n.t('{{amount}} {{token}} Received', { amount, token }), {
-        body: i18n.t('Your account just received {{amount}} {{token}} {{message}}', {
-          amount,
-          token,
-          message,
-        }),
-      });
-      toast.info(
-        i18n.t('Your account just received {{amount}} {{token}} {{message}}', {
-          amount,
-          token,
-          message,
-        })
-      );
-    }
-  });
-};
-
-const TransactionSocketWrapper = () => {
+export const useTransactionUpdate = () => {
   const queryClient = useQueryClient();
   const [currentAccount] = useCurrentAccount();
   const [{ chainID }] = useCurrentApplication();
-  const queryConfig = useTransactionsConfig()();
+  const queryConfig = useTransactionsConfig();
   const transactionQueryKeys = [TRANSACTIONS, chainID, queryConfig];
   // Call useTransactions to initialize the cache
   useTransactions();
@@ -69,7 +39,5 @@ const TransactionSocketWrapper = () => {
     return () => client.socket.off('new.transactions');
   }, []);
 
-  return <div />;
 };
 
-export default TransactionSocketWrapper;
