+++ src/modules/transaction/utils/index.js
@@ -3,3 +3,4 @@ export * from './moduleCommand';
 export * from './transaction';
 export * from './transactionDetailsHelper';
 export * from './removeDuplicateTransactions';
+export * from './showNotificationsForIncomingTransactions';
